refactor: move selectors/count/debug helpers to Scan.cs
